feat: make build incremental and add run script
Rust started to compile every library all the times, so I made the build incremental. I do not know why this has become a problem only now.
There is also a problem with the jetson-inference and jetson-utils libraries, without LD_LIBRARY_PATH the engine does not work.
For this reason i made a run.sh script to remember what to do